app/requirements.txt
Django>=4.2
gunicorn
python-dotenv
psycopg[binary]
djangorestframework
